Add suspend/hibernate to shutdown-helper
Upower 0.99 dropped support for suspend/hibernate as an obsolete
feature in this commit:
http://cgit.freedesktop.org/upower/commit/?id=1ee642e705a63f5ad56a6b55e4bf5c4a64c959b8
Add that functionality into the shutdown-helper so these features
will work again. Additionally Upower is an optional build dependency
and the minimum required is 0.9.7 to support FreeBSD 9 and 10.
Thanks to Guido Berhoerster for pointing out updated ways to handle
suspend/hibernate on FreeBSD and OpenBSD.